+++ src/index.ts
@@ -3,5 +3,6 @@ export * from "./interfaces";
 
 export { runTaskList, runTaskListOnce } from "./main";
 export { run, runOnce } from "./runner";
+export { Logger, LogFunctionFactory, consoleLogFactory } from "./logger";
 
 export { getTasks };